Change default bash to shell
The buildkite agent uses the "/bin/bash" as default shell. On some
reduced environments like embedded linux or alpine. They don't have
bash in environments. The hooks will not work. This commit change
the default "/bin/bash" to "/bin/sh" to run buildkite hooks. Be careful,
this is the method only for run hooks as "/bin/sh".